Changed the change board color function so that the cursor color also changes to enhance cursor visibility
Added a new global variable that holds the new cursor color. Every time button is clicked to navigate the menu, when the program creates the cursor, it makes the cursor color the color that is held in the variable. The variable is changed when the board color is changed as the inverse color of the board.